Move import logic to json endpoint
